doc: Remove «All rights reserved»
This has no legal weight, and was needed due to the Buenos Aires
Convention requiring it. Since all members of the Buenos Aires
Convention are now part of the Berne Convention, there is no need
anymore for this phrasing. And while the original statement must not
be touched, we do not need to replicate this in the COPYING file when
it is legally pointless.